Rename store folder for redux in molecules/Filters
